Refactor logic for creating Parquet file metadata
